feat: replace @PathVariable with @RequestParam
