+++ src/main/java/br/com/conectabyte/profissu/controllers/ConversationController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import br.com.conectabyte.profissu.dtos.request.ConversationRequestDto;
@@ -49,6 +50,18 @@ public class ConversationController {
   @GetMapping("/{id}/messages")
   public Page<MessageResponseDto> listMessages(@PathVariable Long id, @ParameterObject Pageable pageable) {
     return this.conversationService.listMessages(id, pageable);
+  }
+
+  @Operation(summary = "Retrieve conversations by user ID", description = "Fetches a paginated list of conversations where the specified user is the requester.", responses = {
+      @ApiResponse(responseCode = "200", description = "Successfully retrieved conversations", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Page.class))),
+      @ApiResponse(responseCode = "400", description = "Invalid pagination parameters", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ExceptionDto.class))),
+      @ApiResponse(responseCode = "401", description = "Invalid or missing authentication credentials", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ExceptionDto.class)))
+  })
+  @GetMapping
+  @PreAuthorize("@securityService.isOwner(#userId) || @securityService.isAdmin()")
+  public Page<ConversationResponseDto> findConversationByUserId(@RequestParam Long userId,
+      @ParameterObject Pageable pageable) {
+    return conversationService.findByUserId(userId, pageable);
   }
 
   @Operation(summary = "Make an offer for a requested service", description = "Allows a user to make an offer by opening a conversation related to a requested service.")
