+++ src/main/java/br/com/conectabyte/profissu/controllers/UserController.java
@@ -1,8 +1,5 @@
 package br.com.conectabyte.profissu.controllers;
 
-import org.springdoc.core.annotations.ParameterObject;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -16,12 +13,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import br.com.conectabyte.profissu.dtos.request.PasswordRequestDto;
 import br.com.conectabyte.profissu.dtos.request.ProfileRequestDto;
-import br.com.conectabyte.profissu.dtos.response.ConversationResponseDto;
 import br.com.conectabyte.profissu.dtos.response.ExceptionDto;
-import br.com.conectabyte.profissu.dtos.response.RequestedServiceResponseDto;
 import br.com.conectabyte.profissu.dtos.response.UserResponseDto;
-import br.com.conectabyte.profissu.services.ConversationService;
-import br.com.conectabyte.profissu.services.RequestedServiceService;
 import br.com.conectabyte.profissu.services.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -37,8 +30,6 @@ import lombok.RequiredArgsConstructor;
 @Tag(name = "Users", description = "Operations related to managing users")
 public class UserController {
   private final UserService userService;
-  private final RequestedServiceService requestedServiceService;
-  private final ConversationService conversationService;
 
   @Operation(summary = "Retrieve user by ID", description = "Fetches a user's details using the provided ID. Requires authentication.", responses = {
       @ApiResponse(responseCode = "200", description = "User successfully retrieved", content = @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponseDto.class))),
@@ -50,29 +41,6 @@ public class UserController {
   @GetMapping("/{id}")
   public ResponseEntity<UserResponseDto> findById(@PathVariable Long id) {
     return ResponseEntity.ok().body(this.userService.findByIdAndReturnDto(id));
-  }
-
-  @Operation(summary = "Retrieve requested services by user ID", description = "Fetches a paginated list of requested services associated with the provided user ID.", responses = {
-      @ApiResponse(responseCode = "200", description = "Successfully retrieved requested services", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Page.class))),
-      @ApiResponse(responseCode = "400", description = "Invalid pagination parameters", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ExceptionDto.class))),
-      @ApiResponse(responseCode = "401", description = "Invalid or missing authentication credentials", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ExceptionDto.class)))
-  })
-  @GetMapping("/{id}/requested-services")
-  public Page<RequestedServiceResponseDto> findRequestedServiceByUserId(@PathVariable Long id,
-      @ParameterObject Pageable pageable) {
-    return requestedServiceService.findByUserId(id, pageable);
-  }
-
-  @Operation(summary = "Retrieve conversations by user ID", description = "Fetches a paginated list of conversations where the specified user is the requester.", responses = {
-      @ApiResponse(responseCode = "200", description = "Successfully retrieved conversations", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Page.class))),
-      @ApiResponse(responseCode = "400", description = "Invalid pagination parameters", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ExceptionDto.class))),
-      @ApiResponse(responseCode = "401", description = "Invalid or missing authentication credentials", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ExceptionDto.class)))
-  })
-  @GetMapping("/{id}/conversations")
-  @PreAuthorize("@securityService.isOwner(#id) || @securityService.isAdmin()")
-  public Page<ConversationResponseDto> findConversationByUserId(@PathVariable Long id,
-      @ParameterObject Pageable pageable) {
-    return conversationService.findByUserId(id, pageable);
   }
 
   @Operation(summary = "Soft delete user profile", description = "Marks the user profile as deleted (soft delete) using the provided ID. Requires authentication.", responses = {
